Cleaned up the bot and moved most of the stuff to the bot.helper namespace
